A CLEAN-FORM FIX FOR FACTORIAL FORM
# WHAT
On applying the `cleanAll()` function to any form,  the `clean()` function is called per field in order to reset the field's value to its original value. As seen in the files, the opposite is done as the `originalValue` instead of the current value. This PR corrects this error by assigning the field's value back to it's original value. Only at this point can we say that the form is truly no longer `dirty` :)