2.1.1 Massive perf improvement
It's said that cache invalidation is one of the hardest problems in
computer science, but that's actually not true when you work with
observable data.

This commit is an example: We should only calculate `data()` once, and
if data does not change, not calculate it ever again. This will have a
positive effect in the whole application.