Use `isEqual` for `isDirty` getter in Field (#40)
This commit fixes an issue where the `isDirty` getter in Field was giving false negatives.

In javascript world:
```ts
console.log([1] === [1]) // `false`
```
This is because we don't check the values of the array, but its allocation in memory. Same happens with objects.

Even `factorial-form` doesn't officially support arrays or objects as fields, this change will luckily solve some small bugs we have in our frontend app (we already detected one).